% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{boot_escapement_plots}
\alias{boot_escapement_plots}
\title{Plot bootstrapped annual estimates of salmon escapement}
\usage{
boot_escapement_plots()
}
\value{

}
\description{
Plot bootstrapped annual estimates of salmon escapement
}
\examples{
\dontrun{

}
}
