% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{daily_passage}
\alias{daily_passage}
\title{Estimate daily passage of salmon based on the top model from \code{model_escapement()}}
\usage{
daily_passage(passage, date_time)
}
\arguments{
\item{passage}{a numeric vector of hourly salmon passage estimates returned from \code{hourly_passage()}}

\item{date_time}{a POSIXct vector of associated dates and times}
}
\value{
a data frame of daily salmon escapement estimates, grouped by day
}
\description{
Estimate daily passage of salmon based on the top model from \code{model_escapement()}
}
\examples{
\dontrun{
daily_passage(passage, dat$date)
}
}
