% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{escapement}
\alias{escapement}
\title{Estimate annual salmon escapement}
\usage{
escapement(passage, date_time)
}
\arguments{
\item{passage}{a numeric vector of hourly salmon escapement estimates returned from \code{hourly_passage()}}

\item{date_time}{a POSIXct vector of associated dates and times}
}
\value{
a data frame of annual salmon escapement estimates, grouped by year
}
\description{
Estimate annual salmon escapement
}
\examples{
\dontrun{
escapement(passage, dat$date)
}
}
