% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{estimate_escapement}
\alias{estimate_escapement}
\title{Estimate annual escapement of salmon based on the top model from \code{model_escapement()}}
\usage{
estimate_escapement(
  input_data = "./data/derived/dat.Rdata",
  model_results = "./output/models/model_results.Rdata",
  saveit = FALSE,
  output
)
}
\arguments{
\item{output}{}
}
\value{

}
\description{
Estimate annual escapement of salmon based on the top model from \code{model_escapement()}
}
\examples{
\dontrun{
estimate_escapement()
}
}
