% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{explore_plots}
\alias{explore_plots}
\title{Create exploratory plot of raw salmon photo counts}
\usage{
explore_plots(input = "./data/derived/dat.Rdata", saveit = FALSE, output)
}
\arguments{
\item{input}{a directory path to an Rdata file containing formatted salmon count data returned by \code{import_format()}}

\item{saveit}{whether to save the returned list of plots as a RData object}

\item{output}{if \code{saveit = TRUE}, the directory path to where to save the returned list of plots}
}
\value{
a list of \code{ggplot2} plots
}
\description{
Creates a list of exploratory plots of raw salmon photo counts. This includes a plot of daily photo counts grouped by year, and a plot of hourly photo counts grouped by year.
}
\examples{
\dontrun{
explore_plots(input = "./data/derived/dat.Rdata", saveit = FALSE)
}
}
