% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{hourly_passage}
\alias{hourly_passage}
\title{Estimate hourly passage of salmon based on the top model from \code{model_escapement()}}
\usage{
hourly_passage(dat, models, saveit = FALSE, output)
}
\arguments{
\item{dat}{a data frame containing formatted salmon count data returned by \code{import_format()}}

\item{models}{a list of model output returned from \code{model_escapement}.}

\item{saveit}{whether to save the results as a RData object}

\item{output}{a numeric vector containing estimated salmon escapement (hourly)}
}
\value{

}
\description{
Estimate hourly passage of salmon based on the top model from \code{model_escapement()}
}
\examples{
\dontrun{
hourly_passage()
}
}
