% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_format.R
\name{import_format}
\alias{import_format}
\title{Import and format photo and video salmon count data}
\usage{
import_format(input, saveit = FALSE, output)
}
\arguments{
\item{input}{the directory path to a CSV containing dates, and hourly video and photo counts of salmon.}

\item{saveit}{whether to save the returned data frame as a RData object}

\item{output}{if \code{saveit = TRUE}, the directory path to where to save the returned data frame}
}
\value{
a data frame of salmon counts formatted for \code{escapement} functions
}
\description{
Import and format photo and video salmon count data
}
\examples{
\dontrun{
import_format("salmon_counts.csv")
}
}
