% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{loadRData}
\alias{loadRData}
\title{An internal function to load and rename an Rdata object}
\usage{
loadRData(input)
}
\arguments{
\item{input}{the directory path to an Rdata object}
}
\value{
a renamed object loaded into R workspace
}
\description{
An internal function to load and rename an Rdata object
}
\examples{
\dontrun{
loadRdata("myfile.Rdata")
}
}
