% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{model_diagnostics}
\alias{model_diagnostics}
\title{Model diagnostics for salmon escapement models from photo and video count data}
\usage{
model_diagnostics(output = "./output/models/model_diagnostics.Rdata")
}
\arguments{
\item{output}{a file directory path to a list of objects model selection results returned by \code{model_selection}}
}
\value{
a list of objects including residual plots, a score test for non-constant error variance (a code{chisqTest} object) returned by \code{car::ncvTest()}, a spread level plot (object of class \code{spreadLevelPlot}) returned by \code{car::spreadLevelPlot()}, influence measures returned by \code{influence.measures()}, and Cook's distances returned by \code{cooks.distance()}
}
\description{
Model diagnostics for salmon escapement models from photo and video count data
}
\examples{
\dontrun{
model_diagnostics(output = "path/to/your/model_diagnostics.Rdata")
}
}
