% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{model_escapement}
\alias{model_escapement}
\title{Model salmon escapement using photo and video counts}
\usage{
model_escapement(dat, saveit = FALSE, output)
}
\arguments{
\item{dat}{a data frame containing formatted salmon count data returned by \code{import_format()}}

\item{saveit}{whether to save the results as a RData object}

\item{output}{if \code{saveit = TRUE}, the directory path to where to save the returned results}
}
\value{
a list containing a list of model output (\code{models}), a data frame of associated AIC values (\code{aic_table}), and the top model based on AIC (\code{top_model})
}
\description{
Model salmon escapement using photo and video counts
}
\examples{
\dontrun{
model_escapement(input = "./data/derived/dat.Rdata")
}
}
