% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_boot_escapement}
\alias{plot_boot_escapement}
\title{Plot bootstrapped annual estimates of salmon escapement}
\usage{
plot_boot_escapement(boots)
}
\arguments{
\item{boots}{a data frame returned by \code{boot_escapement()}}
}
\value{
a \code{ggplot2} Rdata object
}
\description{
Plot bootstrapped annual estimates of salmon escapement
}
\examples{
\dontrun{
plot_boot_escapement(boots)
}
}
