% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_hourly}
\alias{plot_hourly}
\title{Plot estimate hourly salmon escapement by day}
\usage{
plot_hourly(dat, models)
}
\arguments{
\item{dat}{a data frame containing formatted salmon count data returned by \code{import_format()}}

\item{models}{a list of model output returned from \code{model_escapement}}
}
\value{
a \code{ggplot()} facetted plot showing
}
\description{
Plot estimate hourly salmon escapement by day
}
\examples{
\dontrun{
plot_hourly(dat, models)
}
}
