% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_hrly}
\alias{plot_hrly}
\title{Plot estimate hourly salmon escapement by day}
\usage{
plot_hrly(dat)
}
\arguments{
\item{dat}{a data frame containing formatted salmon count data returned by \code{import_format()}}
}
\value{
a \code{ggplot()} object
}
\description{
Plot estimate hourly salmon escapement by day
}
\examples{
\dontrun{
plot_hrly(dat)
}
}
