% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_topmodel}
\alias{plot_topmodel}
\title{Plots of the top model for estimating salmon escapement from photo and video count data}
\usage{
plot_topmodel(models)
}
\arguments{
\item{models}{a list of model output returned from \code{model_escapement}}
}
\value{
a \code{ggplot2} Rdata object
}
\description{
Plots of the top model for estimating salmon escapement from photo and video count data
}
\examples{
\dontrun{
plot_topmodel(models)
}
}
