% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{run_report}
\alias{run_report}
\title{Generate an Rmarkdown report of salmon escapement}
\usage{
run_report(
  outfile = "escapement_report.pdf",
  outdir = getwd(),
  parameters = "ask"
)
}
\arguments{
\item{outfile}{a name of the returned PDF file}

\item{outdir}{a directory path to the folder to save the PDF}
}
\value{
an RMarkdown generated PDF report
}
\description{
Generate an Rmarkdown report of salmon escapement
}
\examples{
\dontrun{
run_report(outfile = "akalura_report.pdf", outdir = "./report")
}
}
