% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{salmon_counts}
\alias{salmon_counts}
\title{Video and photo counts of sockeye salmon at Akalura River, Kodiak Island, Alaska}
\format{
A data frame with columns:
\describe{
 \item{date}{A POSIXlt date}
 \item{photo}{Number of salmon counted using photos (numeric)}
 \item{video}{Number of salmon counted using videos (numeric)}
 \item{julian_day}{The julian day (numeric)}
 \item{year}{The year (numeric)}
}
}
\source{
Kodiak National Wildlife Refuge, Kodiak, Alaska
}
\usage{
salmon_counts
}
\description{
Video and photo counts of sockeye salmon at Akalura River, Kodiak Island, Alaska
}
\examples{
\dontrun{
 salmon_counts
}
}
\keyword{datasets}
