% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{boot_escapement}
\alias{boot_escapement}
\title{Calculate bootstrapped 95% confidence intervals for salmon escapement estimates}
\usage{
boot_escapement()
}
\value{

}
\description{
Calculate bootstrapped 95% confidence intervals for salmon escapement estimates
}
