% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{bootit}
\alias{bootit}
\title{An internal function used to generate bootstrapped confidence intervals around salmon escapement estimates}
\usage{
bootit(year1, dat, models)
}
\arguments{
\item{year1}{a numeric value indicating the year of data to bootstrap}
}
\value{

}
\description{
An internal function used to generate bootstrapped confidence intervals around salmon escapement estimates
}
\examples{
dontrun{
bootit(2015)
}
}
