% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{daily_passage}
\alias{daily_passage}
\title{Estimate daily passage and cumulative daily passage of salmon based on the top model from \code{model_escapement()}}
\usage{
daily_passage(dat, models)
}
\arguments{
\item{dat}{a data frame containing formatted salmon count data returned by \code{import_format()}}

\item{models}{a list of model output returned from \code{model_escapement}}
}
\value{
a data frame containing daily salmon passage and cumulative daily salmon passage
}
\description{
Estimate daily passage and cumulative daily passage of salmon based on the top model from \code{model_escapement()}
}
\examples{
\dontrun{
daily_passage(dat, models)
}
}
