% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{escapement}
\alias{escapement}
\title{Estimate annual salmon escapement}
\usage{
escapement(dat, models)
}
\arguments{
\item{dat}{a data frame containing formatted salmon count data returned by \code{import_format()}}

\item{models}{a list of model output returned from \code{model_escapement}}
}
\value{
a data frame of annual salmon escapement estimates, grouped by year
}
\description{
Estimate annual salmon escapement
}
\examples{
\dontrun{
escapement(passage, dat$date)
}
}
