% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{model_diagnostics}
\alias{model_diagnostics}
\title{Model diagnostics for salmon escapement models from photo and video count data}
\usage{
model_diagnostics(models)
}
\arguments{
\item{models}{a list containing the model output returned from \code{model_escapement}.}
}
\value{
a list of objects including residual plots, a score test for non-constant error variance (a code{chisqTest} object) returned by \code{car::ncvTest()}, a spread level plot (object of class \code{spreadLevelPlot}) returned by \code{car::spreadLevelPlot()}, influence measures returned by \code{influence.measures()}, and Cook's distances returned by \code{cooks.distance()}
}
\description{
Model diagnostics for salmon escapement models from photo and video count data
}
\examples{
\dontrun{
model_diagnostics(models)
}
}
