% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_daily}
\alias{plot_daily}
\title{Plot estimated minimum daily salmon escapement by year}
\usage{
plot_daily(dat)
}
\arguments{
\item{dat}{a data frame containing formatted salmon count data returned by \code{import_format()}}
}
\value{
a \code{ggplot()} object
}
\description{
Plot estimated minimum daily salmon escapement by year
}
\examples{
\dontrun{
plot_daily(dat)
}
}
