% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_daily}
\alias{plot_daily}
\title{Plot estimated minimum daily salmon escapement by year}
\usage{
plot_daily(dat, models)
}
\arguments{
\item{dat}{a data frame containing formatted salmon count data returned by \code{import_format()}}

\item{models}{a list of model output returned from \code{model_escapement}}
}
\value{
a list containign two \code{ggplot()} objects. \code{daily} is a plot of daily salmon passage by year. \code{cumul_daily} is a plot of cumulative salmon passage by year.
}
\description{
Plot estimated minimum daily salmon escapement by year
}
\examples{
\dontrun{
plot_daily(dat, models)
}
}
