% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{run_report}
\alias{run_report}
\title{Generate an Rmarkdown report of Alulara salmon escapement}
\usage{
run_report(
  input_rmd = "akalura_report.rmd",
  outfile = "akalura_report_2018.pdf",
  outdir = "./report"
)
}
\arguments{
\item{input_rmd}{}

\item{outfile}{}

\item{outdir}{}
}
\value{

}
\description{
Generate an Rmarkdown report of Alulara salmon escapement
}
\examples{
\dontrun{
run_report(input_rmd = "akalura_report.rmd", outfile = "akalura_report_2018.pdf", outdir = "./report")
}
}
