% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{run_report}
\alias{run_report}
\title{Generate an Rmarkdown report of salmon escapement}
\usage{
run_report(
  outformat = bookdown::pdf_document2(),
  outfile = "report",
  outdir = getwd(),
  parameters = "ask",
  ...
)
}
\arguments{
\item{outformat}{the file format of the report. Default is \code{bookdown::pdf_document2()}.}

\item{outfile}{a name of the returned file}

\item{outdir}{a directory path to the folder to save the PDF}

\item{parameters}{a list of custom parameters or \code{all} to launch a shiny app to select custom parameters}

\item{...}{pass additional arguments to \code{rmarkdown::render()}}
}
\value{
an RMarkdown-generated report
}
\description{
Generate an Rmarkdown report of salmon escapement
}
\examples{
\dontrun{
run_report(outfile = "akalura_report.pdf", outdir = "./report")
}
}
